requirements.txt
aioprocessing==2.0.1
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
async-class==0.5.0
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
breadability==0.1.20
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
celery==5.4.0
certifi @ file:///C:/b/abs_7bbxxprxir/croot/certifi_1720453609121/work/certifi
cffi==1.16.0
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
docopt==0.6.2
docx==0.2.4
faust-cchardet==2.1.19
filelock @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/filelock_1701807523603/work
fsspec==2024.6.1
gensim==4.3.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
h11==0.14.0
hrequests==0.8.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
intel-openmp==2021.4.0
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
joblib==1.4.2
kombu==5.3.7
llvmlite==0.43.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markupsafe_1707425732791/work
mdurl==0.1.2
mkl==2021.4.0
mkl-fft @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mkl_fft_1699473528480/work
mkl-random @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mkl_random_1699473588250/work
mkl-service==2.4.0
mpmath @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mpmath_1699484863771/work
networkx @ file:///C:/b/abs_36fmumtynt/croot/networkx_1720002497414/work
nltk==3.8.2
numba==0.60.0
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp312-cp312-win_amd64.whl#sha256=becc06674317799ad0165a939a7613809d0bee9bd328a1e4308c57c39cacf08c
openai-whisper==20231117
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parse==1.20.2
pillow @ file:///C:/b/abs_95ls_qh4c2/croot/pillow_1714398870615/work
prompt_toolkit==3.0.47
pycountry==24.6.1
pycparser==2.22
Pygments==2.18.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pyyaml_1699479991712/work
redis==5.0.7
regex==2024.5.15
requests @ file:///C:/b/abs_9c5n24p1y7/croot/requests_1716902867639/work
rich==13.7.1
safetensors==0.4.3
scipy==1.13.1
selectolax==0.3.21
selenium==4.23.0
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
sumy==0.11.0
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.42.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3 @ file:///C:/b/abs_a7hvzm4y95/croot/urllib3_1718912661242/work
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.43.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wrapt==1.16.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
